+++ keyboard.h
@@ -46,7 +46,7 @@ int locWithShifted(const Keyboard *const k, const char c);
 int calcFitnessDirect(Keyboard *k);
 int scoreDigraphDirect(Keyboard *const k, const char digraph[], const int multiplier);
 int calcFitness(Keyboard *k);
-int scoreDigraph(Keyboard *const k, const char digraph[], const int multiplier, const int allLocs[]);
+int scoreDigraph(Keyboard *const k, const char digraph[], const int64_t multiplier, const int allLocs[]);
